o add GameIcon o add soul visual physics feature o fixed bugs according to interest and killed mobs

+++ Assets/Scripts/Tower.cs
@@ -8,8 +8,8 @@ public class Tower {
 	public string towerID;					// tower id, must be unique
 	public string towerName;				// name of the tower
 	public string upgradeID;				// id of the tower, this is going to upgrade to, e.g. 'bullettower2' 
-	public float towerCost;					// price of the Tower to build
-	public float upgradeCost;				// price to upgrade the tower
+	public int towerCost;					// price of the Tower to build
+	public int upgradeCost;				// price to upgrade the tower
 	public int towerDamage;					// damage a shot from this tower does
 	public int towerLevel;					// level of the tower, usually starts at 1
 	public float towerFireRate;				// frequency the tower is shooting
@@ -20,7 +20,7 @@ public class Tower {
 	public Vector2 bulletSize;				// size of projectile
 	public Sprite towerImageHead;			// image of the Towerhead
 
-	public Tower(string _towerID, string _towerName, string _upgradeID, float _towerCost, float _upgradeCost, int _towerDamage, int _towerLevel, float _towerFireRate, float _towerRange, float _aoeRange, float _bulletVelocity, Color32 _towerColor, Vector2 _bulletSize, Sprite _towerImageHead) {
+	public Tower(string _towerID, string _towerName, string _upgradeID, int _towerCost, int _upgradeCost, int _towerDamage, int _towerLevel, float _towerFireRate, float _towerRange, float _aoeRange, float _bulletVelocity, Color32 _towerColor, Vector2 _bulletSize, Sprite _towerImageHead) {
 		towerID = _towerID;
 		towerName = _towerName;
 		upgradeID = _upgradeID;
